Fix three assumptions about type widths and library coverage
All three compile on libstdc++ and none of them on libc++, which is how they were found; two are
wrong everywhere and merely went unnoticed.

CachedSSLContext passed a file_time_type count straight to std::to_string. libc++ represents that
clock in 128 bits, wide enough for any timespec and wider than any to_string overload, so the call
is ambiguous there. Narrowed deliberately: nanoseconds in an int64 still cover 292 years either
side of the epoch, and this is a cache stamp.

RedisConnect parsed a RESP3 double with std::from_chars. libc++ has no floating-point from_chars
and deletes the overload outright. strtod reads the same format, and nothing here touches the
global locale, so it reads the same decimal point. It also initialises the value, which the old
code did not: from_chars leaves its output alone when the text does not parse, and the result was
used either way.

A DateTime test added a time_t to a seconds count under auto. On FreeBSD seconds::rep is the wider
of the two, so the sum deduced long long and gmtime_r had nothing to bind to.

+++ code/src/sputil/net/CachedSSLContext.cpp
@@ -73,7 +73,12 @@ String fileStamp(const filesystem::path& file)
         return "-";
     }
 
-    return to_string(size) + ":" + to_string(written.time_since_epoch().count());
+    // The count is narrowed deliberately. libc++ gives file_time_type a 128-bit representation,
+    // wide enough to hold any timespec, and std::to_string has no overload that takes one - so the
+    // call is ambiguous there while it compiles on libstdc++, whose representation is 64 bits.
+    // Nanoseconds in an int64 still span some 292 years either side of the epoch, which is more
+    // than a cache stamp needs.
+    return to_string(size) + ":" + to_string(static_cast<int64_t>(written.time_since_epoch().count()));
 }
 } // namespace
